refactor: P2-T2 - 增加HTML清洗与URL绝对化提升Markdown质量
